+++ main/src/main/webapp/src/js/services/ActionService.js
@@ -22,6 +22,7 @@ import {ExecuteScriptAction} from "../entities/actions/webActions/ExecuteScriptA
 import {FillWebAction} from "../entities/actions/webActions/FillWebAction";
 import {ClickWebAction} from "../entities/actions/webActions/ClickWebAction";
 import {ClickLinkByTextWebAction} from "../entities/actions/webActions/ClickLinkByTextWebAction";
+import {CheckNodeAttributeValueAction} from "../entities/actions/webActions/CheckAttributeValueAction";
 import {ClearWebAction} from "../entities/actions/webActions/ClearWebAction";
 import {CheckPageTitleAction} from "../entities/actions/webActions/CheckPageTitleAction";
 import {PressKeyAction} from "../entities/actions/webActions/PressKeyAction";
@@ -90,6 +91,8 @@ export class ActionService {
                 return new CheckForNodeWebAction(data);
             case actionType.WEB_PRESS_KEY:
                 return new PressKeyAction(data);
+            case actionType.WEB_CHECK_ATTRIBUTE_VALUE:
+                return new CheckNodeAttributeValueAction(data);
 
             // rest actions
             case actionType.REST_CALL: